requirements.txt
# Yelp Review Theme Discovery - Requirements
# Python 3.10+

# Core Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0

# NLP & Embeddings
sentence-transformers>=2.2.0
spacy>=3.5.0
nltk>=3.8.0
contractions>=0.1.73
beautifulsoup4>=4.12.0
unidecode>=1.3.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM Integration
openai>=1.0.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Development
pytest>=7.0.0
black>=23.0.0
